+++ lib/wechat/core.rb
@@ -6,6 +6,9 @@ require 'wechat/core/access_token'
 require 'wechat/core/server_address'
 require 'wechat/core/tiny_link'
 
+require 'wechat/core/follower'
+require 'wechat/core/follower_profile'
+
 module Wechat
   module Core
     # Your code goes here...