+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # Wechat Core Change Log 微信核心库变更日志
 
 ## v0.1
-
 1. Common module
 2. Access Token wrapper class
 
 ## v0.2
-
 1. Server Address wrapper class
 2. Tiny Link wrapper class
+
+## v0.3
+1. Follower wrapper class
+2. Follower Profile wrapper class